+++ src/CH/shortest_path_algorithms/dijkstra_utils.hpp
@@ -6,12 +6,12 @@
 namespace RouterCH
 {
 typedef std::vector<bool> QSTable; //true jeśli w Q (nie został policzony)
-typedef std::vector<unsigned int> CostTable;
-typedef std::vector<signed int> PathTable;
+typedef std::vector<double> CostTable;
+typedef std::vector<int64_t> PathTable;
 
 unsigned int getIndexOfNextNode(CostTable& costTable, const QSTable& qsTable);
-Route createShortestPath(EdgesTable &edgesTable, const PathTable& pathTable, const unsigned int start,
-                         const unsigned int end, const Nodes& nodes,const ShorctutsTable* shortcutsTable = nullptr);
+Route createShortestPath(EdgesTable &edgesTable, const PathTable& pathTable, const int64_t start,
+                         const int64_t end, const Nodes& nodes,const ShorctutsTable* shortcutsTable = nullptr);
 bool operator==(Route a, Route b);
 bool operator!=(Route a, Route b);
 bool operator==(Node a, Node b);
